+++ packages/sphinx-test-reports/sphinxcontrib/test_reports/directives/test_suite.py
@@ -4,6 +4,7 @@ from docutils import nodes
 from docutils.parsers.rst import Directive, directives
 
 from sphinxcontrib.test_reports.directives import TestCommonDirective
+import sphinxcontrib.test_reports.directives.test_case
 from sphinxcontrib.test_reports.exceptions import TestReportInvalidOption
 from sphinxcontrib.needs.api import add_need
 
@@ -41,7 +42,6 @@ class TestSuiteDirective(TestCommonDirective):
         if suite_name is None:
             raise TestReportInvalidOption('Suite not given!')
 
-
         suite = None
         for suite_obj in self.results:
             if suite_obj['name'] == suite_name:
@@ -66,4 +66,30 @@ class TestSuiteDirective(TestCommonDirective):
                                  status=self.test_status, collapse=self.collapse,
                                  file=self.test_file_given, suite=suite['name'], cases=cases,
                                  passed=passed, skipped=skipped, failed=failed, errors=errors)
+
+        if 'auto_cases' in self.options.keys():
+            for case in suite['testcases']:
+                case_id = self.test_id
+                case_id += '_' + hashlib.sha1(case['classname'].encode("UTF-8") + case['name'].encode("UTF-8")
+                                              ).hexdigest().upper()[:3]
+
+                options = self.options
+                options['case'] = case['name']
+                options['classname'] = case['classname']
+                options['id'] = case_id
+
+                if 'links' not in self.options:
+                    options['links'] = self.test_id
+                elif self.test_id not in options['links']:
+                    options['links'] = options['links'] + ';' + self.test_id
+
+                arguments = [case['name']]
+                case_directive = sphinxcontrib.test_reports.directives.test_case.TestCaseDirective(
+                    'test-case', arguments, options,
+                    self.content, self.lineno,
+                    self.content_offset, self.block_text, self.state,
+                    self.state_machine)
+
+                main_section += case_directive.run()
+
         return main_section